fix: seperate payment history page in to several component and modify server action for more performance
